fix: handler errors in posts

+++ src/posts/posts.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from '@nestjs/common';
 import { CreatePostDto } from './dto/create-post.dto';
 import { PostsRepository } from './posts.respository';
 
@@ -12,11 +12,27 @@ export class PostsService {
   }
 
   async findAll() {
-    return await this.repository.findAll();
+    const posts = await this.repository.findAll();
+    const filteredPosts = posts.map(post => {
+      const { id, title, text, image } = post;
+      return {
+        id,
+        title,
+        text,
+        ...(image !== null && { image })
+      };
+    });
+    return filteredPosts;
   }
 
-  async findOne(id: number) {
-    return await this.repository.findOne(id);
+  async findOne(idPosts: number) {
+    const posts = await this.repository.findOne(idPosts);
+
+    if(!posts) throw new HttpException("Not Found", HttpStatus.NOT_FOUND)
+
+    const { id, title, text, image } = posts;
+
+    return {id, title, text, ...(image !== null && { image })}
   }
 
   async update(id: number, updatePostDto: CreatePostDto) {
@@ -24,6 +40,10 @@ export class PostsService {
   }
 
   async remove(id: number) {
+    const publications = await this.repository.findMedia(id);
+    
+    if(publications) throw new HttpException('Forbidden', HttpStatus.FORBIDDEN);
+
     return await this.repository.remove(id);
   }
 }
